+++ packages/box-core/src/tests/onchange.test.js
@@ -3,12 +3,12 @@ import 'jsdom-global/register'
 import React from 'react';
 import { Provider } from 'react-redux';
 import Box from '../index'
-import { mount } from 'enzyme';
+import { render, fireEvent } from '@testing-library/react';
 import { combineReducers, createStore } from 'redux';
 import { userSlice } from './test-utils';
 
 const Text = ({ text, value }) => <h1>{text || value}</h1>
-const Input = ({ onChange, value = '', name }) => <input name={name} value={value} onChange={evt => onChange(evt.target.value)} />
+const Input = ({ onChange, value = '', name, childId }) => <input data-testid={childId} name={name} value={value} onChange={evt => onChange(evt.target.value)} />
 const Container = ({ children }) => <div>{children}</div>
 
 Box.setComponents({
@@ -28,7 +28,7 @@ describe('Test children update onChange', () => {
     );
   });
 
-  it('Quando passo un valore ad un input valido, mi aspetto il render dell\'elemento correttamente', () => {
+  it('Quando passo un valore ad un input valido, mi aspetto il render dell\'elemento correttamente', async () => {
 
     const children = [
       {
@@ -40,7 +40,7 @@ describe('Test children update onChange', () => {
 
     const dispatchSpy = jest.spyOn(store, 'dispatch')
 
-    const wrapper = mount(
+    const wrapper = render(
       <Provider store={store}>
         <div>
           <Box data={children} prefix="app" />
@@ -48,14 +48,15 @@ describe('Test children update onChange', () => {
       </Provider>
     );
 
-    wrapper.find('input').simulate('change', { target: { value: 'Hello' } });
-    expect(dispatchSpy.mock.lastCall[0].type).toEqual('@box/app/update');
+    const inputEl = await wrapper.getByTestId('app.surname')
+    fireEvent.change(inputEl, { target: { value: 'Hello' } })
 
-    expect(wrapper.html()).toEqual('<div><input name="surname" value=\"Hello\"></div>')
+    expect(dispatchSpy.mock.lastCall[0].type).toEqual('@box/app/update');
+    expect(wrapper.baseElement.innerHTML).toEqual('<div><div><input data-testid="app.surname" name="surname" value=\"Hello\"></div></div>')
 
   })
 
-  it('Quando passo un valore ad un input valido complesso e un valore di default, mi aspetto il render dell\'elemento correttamente', () => {
+  it('Quando passo un valore ad un input valido complesso e un valore di default, mi aspetto il render dell\'elemento correttamente', async () => {
 
     const children = [
       {
@@ -72,7 +73,7 @@ describe('Test children update onChange', () => {
 
     const dispatchSpy = jest.spyOn(store, 'dispatch')
 
-    const wrapper = mount(
+    const wrapper = render(
       <Provider store={store}>
         <div>
           <Box data={children} prefix="app" />
@@ -80,7 +81,10 @@ describe('Test children update onChange', () => {
       </Provider>
     );
 
-    wrapper.find('input[name="surname"]').simulate('change', { target: { value: 'Hello' } });
+
+    const inputEl = await wrapper.getByTestId('app.a.b.c.surname')
+    fireEvent.change(inputEl, { target: { value: 'Hello' } })
+    
     expect(dispatchSpy.mock.lastCall[0].type).toEqual('@box/app/update');
 
     expect(store.getState()).toEqual({
@@ -96,11 +100,11 @@ describe('Test children update onChange', () => {
       }
     })
 
-    expect(wrapper.html()).toEqual('<div><input name="surname" value=\"Hello\"><input value=\"andrea\"></div>')
+    expect(wrapper.baseElement.innerHTML).toEqual('<div><div><input data-testid=\"app.a.b.c.surname\" name=\"surname\" value=\"Hello\"><input data-testid=\"app.a.name\" value=\"andrea\"></div></div>')
 
   })
 
-  it('Quando passo un valore ad un input obbligatorio, mi aspetto che sia validato (^isValid)', () => {
+  it('Quando passo un valore ad un input obbligatorio, mi aspetto che sia validato (^isValid)', async () => {
 
     const children = [
       {
@@ -120,7 +124,7 @@ describe('Test children update onChange', () => {
 
     const dispatchSpy = jest.spyOn(store, 'dispatch')
 
-    const wrapper = mount(
+    const wrapper = render(
       <Provider store={store}>
         <div>
           <Box data={children} prefix="app" />
@@ -129,10 +133,12 @@ describe('Test children update onChange', () => {
     );
 
 
-    wrapper.find('input').simulate('change', { target: { value: 'Hello' } });
+    const inputEl = await wrapper.getByTestId('app.surname')
+    fireEvent.change(inputEl, { target: { value: 'Hello' } })
+
     expect(dispatchSpy.mock.lastCall[0].type).toEqual('@box/app/update');
 
-    expect(wrapper.html()).toEqual('<div><input name="surname" value=\"Hello\"><h1>OKK</h1></div>')
+    expect(wrapper.baseElement.innerHTML).toEqual('<div><div><input data-testid=\"app.surname\" name=\"surname\" value=\"Hello\"><h1>OKK</h1></div></div>')
 
   })
 
@@ -161,7 +167,7 @@ describe('Test children update onChange', () => {
       }
     ]
 
-    const wrapper = mount(
+    const wrapper = render(
       <Provider store={store}>
         <div>
           <Box data={children} prefix="app" />
@@ -169,7 +175,7 @@ describe('Test children update onChange', () => {
       </Provider>
     );
 
-    expect(wrapper.html()).toEqual('<div><input name="surname" value=\"\"><h1>KO!</h1></div>')
+    expect(wrapper.baseElement.innerHTML).toEqual('<div><div><input data-testid=\"app.surname\" name=\"surname\" value=\"\"><h1>KO!</h1></div></div>')
 
   })