+++ packages/box-core/src/tests/setup.js
@@ -1,10 +1,2 @@
-import Enzyme from 'enzyme'
-import Adapter from 'enzyme-adapter-react-16'
-// import Adapter from '@wojtekmaj/enzyme-adapter-react-17' // Adapter ufficiale di enzyme per React 17 non ancora disponibile
-
+global.IS_REACT_ACT_ENVIRONMENT = true
 console.warn = function() {}
-// console.log = function() {}
-
-Enzyme.configure({
-  adapter: new Adapter()
-})