+++ packages/box-core/changelog.md
@@ -5,6 +5,7 @@
 
 ### 5.0.0 
 - Rename props fields in children
+- Rename props field in child
 - Rename extendControls in registerComponents
 - Rename _fields in _children
 - Rename renderFields in renderChildren